+++ modules/apps/headless-apio/blog/blog-apio-impl/src/main/java/com/liferay/blog/apio/internal/architect/form/BlogPostingForm.java
@@ -15,9 +15,9 @@
 package com.liferay.blog.apio.internal.architect.form;
 
 import com.liferay.apio.architect.form.Form;
-import com.liferay.apio.architect.form.Form.Builder;
 import com.liferay.apio.architect.function.throwable.ThrowableFunction;
 import com.liferay.apio.architect.functional.Try;
+import com.liferay.category.apio.architect.identifier.CategoryIdentifier;
 import com.liferay.media.object.apio.architect.identifier.MediaObjectIdentifier;
 import com.liferay.person.apio.architect.identifier.PersonIdentifier;
 import com.liferay.portal.kernel.repository.model.FileEntry;
@@ -51,7 +51,7 @@ public class BlogPostingForm {
 	 * @review
 	 */
 	public static Form<BlogPostingForm> buildForm(
-		Builder<BlogPostingForm> formBuilder) {
+		Form.Builder<BlogPostingForm> formBuilder) {
 
 		return formBuilder.title(
 			__ -> "The blog posting form"
@@ -69,6 +69,8 @@ public class BlogPostingForm {
 			"creator", PersonIdentifier.class, BlogPostingForm::setCreatorId
 		).addOptionalLinkedModel(
 			"image", MediaObjectIdentifier.class, BlogPostingForm::setImageId
+		).addOptionalLinkedModelList(
+			"category", CategoryIdentifier.class, BlogPostingForm::setCategories
 		).addOptionalString(
 			"alternativeHeadline", BlogPostingForm::setAlternativeHeadline
 		).addOptionalString(
@@ -259,6 +261,11 @@ public class BlogPostingForm {
 			serviceContext.setAssetTagNames(ArrayUtil.toStringArray(_keywords));
 		}
 
+		if (ListUtil.isNotEmpty(_categories)) {
+			serviceContext.setAssetCategoryIds(
+				ArrayUtil.toLongArray(_categories));
+		}
+
 		return serviceContext;
 	}
 
@@ -268,6 +275,10 @@ public class BlogPostingForm {
 
 	public void setArticleBody(String articleBody) {
 		_articleBody = articleBody;
+	}
+
+	public void setCategories(List<Long> categories) {
+		_categories = categories;
 	}
 
 	public void setCreateDate(Date createDate) {
@@ -312,6 +323,7 @@ public class BlogPostingForm {
 
 	private String _alternativeHeadline;
 	private String _articleBody;
+	private List<Long> _categories;
 	private Date _createDate;
 	private Long _creatorId;
 	private String _description;