+++ modules/apps/headless-apio/media-object/media-object-apio-impl/src/main/java/com/liferay/media/object/apio/internal/architect/resource/MediaObjectNestedCollectionResource.java
@@ -16,9 +16,17 @@ package com.liferay.media.object.apio.internal.architect.resource;
 
 import static com.liferay.portal.apio.idempotent.Idempotent.idempotent;
 
+import com.liferay.adaptive.media.AMAttribute;
+import com.liferay.adaptive.media.AdaptiveMedia;
+import com.liferay.adaptive.media.image.finder.AMImageFinder;
+import com.liferay.adaptive.media.image.finder.AMImageQueryBuilder;
+import com.liferay.adaptive.media.image.mime.type.AMImageMimeTypeProvider;
+import com.liferay.adaptive.media.image.processor.AMImageAttribute;
+import com.liferay.adaptive.media.image.processor.AMImageProcessor;
 import com.liferay.apio.architect.functional.Try;
 import com.liferay.apio.architect.pagination.PageItems;
 import com.liferay.apio.architect.pagination.Pagination;
+import com.liferay.apio.architect.representor.NestedRepresentor;
 import com.liferay.apio.architect.representor.Representor;
 import com.liferay.apio.architect.resource.NestedCollectionResource;
 import com.liferay.apio.architect.routes.ItemRoutes;
@@ -41,6 +49,8 @@ import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.util.ListUtil;
 
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -72,7 +82,7 @@ public class MediaObjectNestedCollectionResource
 
 	@Override
 	public String getName() {
-		return "media-object";
+		return "document";
 	}
 
 	@Override
@@ -106,16 +116,17 @@ public class MediaObjectNestedCollectionResource
 			"dateModified", FileEntry::getModifiedDate
 		).addLinkedModel(
 			"creator", PersonIdentifier.class, FileEntry::getUserId
+		).addNestedList(
+			"adaptedMedia", this::_getAdaptiveMedias,
+			this::_getAdaptiveMediaNestedRepresentor
 		).addNumber(
 			"sizeInBytes", FileEntry::getSize
 		).addRelatedCollection(
 			"category", CategoryIdentifier.class
 		).addString(
-			"contentSize", fileEntry -> String.valueOf(fileEntry.getSize())
-		).addString(
 			"description", FileEntry::getDescription
 		).addString(
-			"fileFormat", FileEntry::getMimeType
+			"encodingFormat", FileEntry::getMimeType
 		).addString(
 			"headline", FileEntry::getTitle
 		).addString(
@@ -123,6 +134,66 @@ public class MediaObjectNestedCollectionResource
 		).addStringList(
 			"keywords", this::_getMediaObjectAssetTags
 		).build();
+	}
+
+	private NestedRepresentor<AdaptiveMedia<AMImageProcessor>>
+		_getAdaptiveMediaNestedRepresentor(
+			NestedRepresentor.Builder<AdaptiveMedia<AMImageProcessor>>
+				builder) {
+
+		return builder.types(
+			"ImageObject", "MediaObject"
+		).addNumber(
+			"height",
+			adaptiveMedia -> _getAdaptiveMediaValue(
+				adaptiveMedia, AMImageAttribute.AM_IMAGE_ATTRIBUTE_HEIGHT)
+		).addNumber(
+			"sizeInBytes",
+			adaptiveMedia -> _getAdaptiveMediaValue(
+				adaptiveMedia, AMAttribute.getContentLengthAMAttribute())
+		).addNumber(
+			"width",
+			adaptiveMedia -> _getAdaptiveMediaValue(
+				adaptiveMedia, AMImageAttribute.AM_IMAGE_ATTRIBUTE_WIDTH)
+		).addRelativeURL(
+			"contentUrl",
+			adaptiveMedia -> String.valueOf(adaptiveMedia.getURI())
+		).addString(
+			"resolutionName",
+			adaptiveMedia -> _getAdaptiveMediaValue(
+				adaptiveMedia, AMAttribute.getConfigurationUuidAMAttribute())
+		).build();
+	}
+
+	private List<AdaptiveMedia<AMImageProcessor>> _getAdaptiveMedias(
+		FileEntry fileEntry) {
+
+		return Try.fromFallible(
+			fileEntry::getMimeType
+		).filter(
+			_amImageMimeTypeProvider::isMimeTypeSupported
+		).map(
+			mimeType -> _amImageFinder.getAdaptiveMediaStream(
+				amImageQueryBuilder -> amImageQueryBuilder.forFileEntry(
+					fileEntry
+				).withConfigurationStatus(
+					AMImageQueryBuilder.ConfigurationStatus.ANY
+				).done()
+			).collect(
+				Collectors.toList()
+			)
+		).orElse(
+			null
+		);
+	}
+
+	private <V> V _getAdaptiveMediaValue(
+		AdaptiveMedia<AMImageProcessor> adaptiveMedia,
+		AMAttribute<AMImageProcessor, V> amAttribute) {
+
+		Optional<V> valueOptional = adaptiveMedia.getValueOptional(amAttribute);
+
+		return valueOptional.orElse(null);
 	}
 
 	private FileEntry _getFileEntry(
@@ -162,6 +233,12 @@ public class MediaObjectNestedCollectionResource
 
 		return new PageItems<>(fileEntries, count);
 	}
+
+	@Reference
+	private AMImageFinder _amImageFinder;
+
+	@Reference
+	private AMImageMimeTypeProvider _amImageMimeTypeProvider;
 
 	@Reference
 	private AssetTagLocalService _assetTagLocalService;