+++ portal-impl/src/com/liferay/portal/language/LanguageImpl.java
@@ -72,6 +72,7 @@ import java.util.Map;
 import java.util.ResourceBundle;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.function.Supplier;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -1067,6 +1068,17 @@ public class LanguageImpl implements Language, Serializable {
 
 	@Override
 	public Locale getLocale(long groupId, String languageCode) {
+		try {
+			if (isInheritLocales(groupId)) {
+				return getLocale(languageCode);
+			}
+		}
+		catch (Exception e) {
+			if (_log.isWarnEnabled()) {
+				_log.warn("Unable to check if group inherits locales");
+			}
+		}
+
 		Map<String, Locale> groupLanguageCodeLocalesMap =
 			_getGroupLanguageCodeLocalesMap(groupId);
 
@@ -1537,11 +1549,26 @@ public class LanguageImpl implements Language, Serializable {
 		return language1.equals(language2);
 	}
 
+	/**
+	 * @deprecated As of Judson (7.1.x), replaced by {@link #process(
+	 *            Supplier, Locale, String)}
+	 */
+	@Deprecated
 	@Override
 	public String process(
 		ResourceBundle resourceBundle, Locale locale, String content) {
 
-		StringBundler sb = new StringBundler();
+		return process(() -> resourceBundle, locale, content);
+	}
+
+	@Override
+	public String process(
+		Supplier<ResourceBundle> resourceBundleSupplier, Locale locale,
+		String content) {
+
+		StringBundler sb = null;
+
+		ResourceBundle resourceBundle = null;
 
 		Matcher matcher = _pattern.matcher(content);
 
@@ -1552,8 +1579,16 @@ public class LanguageImpl implements Language, Serializable {
 
 			String key = matcher.group(1);
 
+			if (sb == null) {
+				sb = new StringBundler();
+			}
+
 			sb.append(content.substring(x, y));
 			sb.append(StringPool.APOSTROPHE);
+
+			if (resourceBundle == null) {
+				resourceBundle = resourceBundleSupplier.get();
+			}
 
 			String value = get(resourceBundle, key);
 
@@ -1562,6 +1597,10 @@ public class LanguageImpl implements Language, Serializable {
 			sb.append(StringPool.APOSTROPHE);
 
 			x = matcher.end(0);
+		}
+
+		if (sb == null) {
+			return content;
 		}
 
 		sb.append(content.substring(x));
@@ -1625,8 +1664,12 @@ public class LanguageImpl implements Language, Serializable {
 
 		String[] languageIds = PropsValues.LOCALES_ENABLED;
 
+		Locale defaultLocale = LocaleUtil.getDefault();
+
 		try {
 			Group group = GroupLocalServiceUtil.getGroup(groupId);
+
+			defaultLocale = PortalUtil.getSiteDefaultLocale(group);
 
 			UnicodeProperties typeSettingsProperties =
 				group.getTypeSettingsProperties();
@@ -1644,6 +1687,9 @@ public class LanguageImpl implements Language, Serializable {
 		HashMap<String, Locale> groupLanguageCodeLocalesMap = new HashMap<>();
 		HashMap<String, Locale> groupLanguageIdLocalesMap =
 			new LinkedHashMap<>();
+
+		groupLanguageCodeLocalesMap.put(
+			defaultLocale.getLanguage(), defaultLocale);
 
 		for (String languageId : languageIds) {
 			Locale locale = LocaleUtil.fromLanguageId(languageId, false);
@@ -1939,6 +1985,17 @@ public class LanguageImpl implements Language, Serializable {
 					languageIds = PropsValues.LOCALES_ENABLED;
 				}
 			}
+
+			Locale defaultLocale = LocaleUtil.getDefault();
+
+			String defaultLanguageId = LocaleUtil.toLanguageId(defaultLocale);
+
+			_languageCodeLocalesMap.put(
+				defaultLocale.getLanguage(), defaultLocale);
+
+			_languageIdLocalesMap.put(defaultLanguageId, defaultLocale);
+
+			languageIds = ArrayUtil.remove(languageIds, defaultLanguageId);
 
 			Set<String> duplicateLanguageCodes = new HashSet<>();